refactor: move and rename proposal counter service
