+++ app/shared/counters/proposal-counter.server.ts
