+++ tests/factories/proposals.ts
@@ -11,7 +11,7 @@ import type {
 import { ConfirmationStatus, DeliberationStatus, PublicationStatus, TalkLevel } from 'prisma/generated/client.ts';
 import type { ProposalCreateInput } from 'prisma/generated/models.ts';
 import { EventSpeakerForProposal } from '~/features/event-participation/speaker-proposals/services/event-speaker-for-proposal.ts';
-import { getNextProposalNumber } from '~/shared/utils/proposal-number.server.ts';
+import { getNextProposalNumber } from '~/shared/counters/proposal-counter.server.ts';
 import { db } from '../../prisma/db.server.ts';
 import { applyTraits } from './helpers/traits.ts';
